+++ pollen_vision/tests/test_config.py
@@ -8,7 +8,7 @@ from pollen_vision.camera_wrappers.depthai.utils import (
     get_config_file_path,
     get_config_files_names,
 )
-from pollen_vision.vision_models.utils import get_checkpoint_path, get_checkpoints_names
+from pollen_vision.perception.utils import get_checkpoint_path, get_checkpoints_names
 
 
 def test_config() -> None:
