+++ scripts/annotate_video.py
@@ -23,9 +23,9 @@ import asyncio
 import cv2
 import numpy as np
 import tqdm
+from pollen_vision.perception.utils import Annotator, get_bboxes
 from pollen_vision.vision_models.object_detection import OwlVitWrapper
 from pollen_vision.vision_models.object_segmentation import MobileSamWrapper
-from pollen_vision.vision_models.utils import Annotator, get_bboxes
 from recorder import Recorder
 
 argParser = argparse.ArgumentParser(description="record sr")
