+++ scripts/test_yolo_world.py
@@ -1,9 +1,8 @@
 import cv2
-import numpy as np
 from pollen_vision.camera_wrappers.depthai import SDKWrapper
 from pollen_vision.camera_wrappers.depthai.utils import get_config_file_path
+from pollen_vision.perception.utils import Annotator
 from pollen_vision.vision_models.object_detection import YoloWorldWrapper
-from pollen_vision.vision_models.utils import Annotator
 
 w = SDKWrapper(get_config_file_path("CONFIG_SR"))
 A = Annotator()
